7.3.7 GA8 release

+++ modules/apps/commerce/commerce-price-list-service/src/main/java/com/liferay/commerce/price/list/internal/upgrade/v1_1_0/CommercePriceEntryUpgradeProcess.java
@@ -25,13 +25,11 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.ObjectValuePair;
 
-import java.sql.DatabaseMetaData;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Statement;
 
 import java.util.List;
-import java.util.Objects;
 
 /**
  * @author Alec Sloan
@@ -87,7 +85,9 @@ public class CommercePriceEntryUpgradeProcess
 			}
 		}
 
-		runSQL("drop index IX_2083879C on CommercePriceEntry");
+		if (hasIndex("CommercePriceEntry", "IX_2083879C")) {
+			runSQL("drop index IX_2083879C on CommercePriceEntry");
+		}
 
 		runSQL("alter table CommercePriceEntry drop column CPInstanceId");
 	}
@@ -108,7 +108,7 @@ public class CommercePriceEntryUpgradeProcess
 						indexMetadata.getIndexName(), tableName));
 			}
 
-			if (!_tableHasIndex(tableName, indexMetadata.getIndexName())) {
+			if (!hasIndex(tableName, indexMetadata.getIndexName())) {
 				runSQL(indexMetadata.getCreateSQL(null));
 			}
 			else if (_log.isInfoEnabled()) {
@@ -118,28 +118,6 @@ public class CommercePriceEntryUpgradeProcess
 						indexMetadata.getIndexName(), tableName));
 			}
 		}
-	}
-
-	private boolean _tableHasIndex(String tableName, String indexName)
-		throws Exception {
-
-		DatabaseMetaData metadata = connection.getMetaData();
-
-		try (ResultSet rs = metadata.getIndexInfo(
-				null, null, tableName, false, false)) {
-
-			while (rs.next()) {
-				String curIndexName = rs.getString("index_name");
-
-				if (Objects.equals(indexName, curIndexName)) {
-					return true;
-				}
-			}
-		}
-		catch (Exception exception) {
-		}
-
-		return false;
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
