+++ modules/apps/knowledge-base/knowledge-base-service/src/main/java/com/liferay/knowledge/base/internal/upgrade/v1_3_1/UpgradeKBComment.java
@@ -27,7 +27,9 @@ public class UpgradeKBComment extends UpgradeProcess {
 	}
 
 	protected void upgradeIndexes() throws Exception {
-		runSQL("drop index IX_FD56A55D on KBComment");
+		if (hasIndex("KBComment", "IX_FD56A55D")) {
+			runSQL("drop index IX_FD56A55D on KBComment");
+		}
 
 		runSQL(
 			"create index IX_FD56A55D on KBComment(userId, classNameId, " +
